fix: 검색완료시에도  검색되게  수정 및 메인페이지에서 에러 수정
